% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCD_fun.R
\name{compare_HMD}
\alias{compare_HMD}
\title{compare HCD aggregate data with HMD}
\usage{
compare_HMD(country)
}
\arguments{
\item{country}{character string with HCD country code}
}
\description{
write stuff here describing what the function does.
}
\details{
write tips here, stuff to watch out for
}
